+++ packages/editor/src/markup/codemirror/line-highlight/index.ts
@@ -1,2 +1,3 @@
-export type {LineRange, LineHighlightOptions} from './extension';
+export type {LineRange} from './types';
+export type {LineHighlightOptions} from './extension';
 export {lineHighlight, setHighlightedLine} from './extension';
